Uploading new GUI
This .zip file contains a new version of the game that incorporates a GUI. GUI.java works with Game.java to allow user input through the GUI to be properly handled. This version is primarily to evaluate the feasibility of this particular design approach. If we agree that it will work, then it can become the main version and we can focus on polishing the interface and exterminating the remaining bugs.

Note: This is still very rough. There are multiple bugs present including, but not limited to: the Raise option being seen as invalid, possible issue with blinds being posted twice by each player, no output is being sent to the textArea. Again, this is very rough, and that is why I segregated it in its own little .zip file. I suggest unpacking this into a new Java project, so as to not disturb your other local version. Also, the CardPics file must be included in its entirety and dropped right into the main project. It should be an immediate subfolder of the main project directory (not a subf