GUI Updates
I updated the GUI to improve its look. Now, the layout is nice and straight with consistent spacing between elements. The method showUserStatistics was added to game at line 559 and is called on line 528. This method displays relevant information about the user's current hand. Set multiple methods in Game to private for clarify and to prevent outside access. Added input checking to multiple methods in Player. Adjusted certain methods elsewhere to improve function and clarify.