bug fixes
Fixed a couple of bugs.  Ran into an issue where the bot was trying to raise more it could (happened when the bot was responding to a raise). In this scenario I modified the bot raise amount to essentially go all in.  We also had a bug which prevented the user from calling an all-in with the all-in button.  basically just changed a couple > to >= to accommodate the scenario.