Added Multiple Games
Modified Game and Driver so now if the player wins, loses, or quits a game they will be asked if they want to play another game. This is handled by the new playGames method in Game, which is now called in Driver. A resetGame method has been added to reset the game state so that data such as chip counts cannot persist from game to game like they do from hand to hand.